+++ Sources/SwiftLinkPreview.swift
@@ -556,7 +556,7 @@ extension SwiftLinkPreview {
                             let value = value.decoded.extendedTrim
                             if tag == "image" {
                                 let value = addImagePrefixIfNeeded(value, result: result)
-                                if value.isImage() { result.set(value, for: key) }
+                                if value.isOpenGraphImage(){ result.set(value, for: key) }
                             } else if tag == "video" {
                                 let value = addImagePrefixIfNeeded(value, result: result)
                                 if value.isVideo() { result.set(value, for: key) }
@@ -621,23 +621,25 @@ extension SwiftLinkPreview {
             let images = result.images
 
             if images == nil || images?.isEmpty ?? true {
-                let values = Regex.pregMatchAll(htmlCode, regex: Regex.imageTagPattern, index: 2)
-                if !values.isEmpty {
-                    let imgs = values.map { self.addImagePrefixIfNeeded($0, result: result) }
 
-                    result.images = imgs
-                    result.image = imgs.first
-                }
-                else{
-                    let values = Regex.pregMatchAll(htmlCode, regex: Regex.secondaryImageTagPattern, index: 1)
+                // Should look for <meta property="og:image" content=""/> first instead of <img/> tag.
+                let values = Regex.pregMatchAll(htmlCode, regex: Regex.secondaryImageTagPattern, index: 2)
+                if !values.isEmpty {
+                    result.images = values
+                    result.image = values.first
+                } else {
+                    // If no OpenGraph image found pick any from <img/> tag to show.
+                    let values = Regex.pregMatchAll(htmlCode, regex: Regex.imageTagPattern, index: 2)
                     if !values.isEmpty {
-                        result.images = values
-                        result.image = values.first
+                        let imgs = values.map { self.addImagePrefixIfNeeded($0, result: result) }
+                        result.images = imgs
+                        result.image = imgs.first
                     }
                 }
+
             }
         } else {
-                let values = Regex.pregMatchAll(htmlCode, regex: Regex.secondaryImageTagPattern, index: 1)
+                let values = Regex.pregMatchAll(htmlCode, regex: Regex.secondaryImageTagPattern, index: 2)
                 if !values.isEmpty {
                     result.images = values
                     result.image = values.first
@@ -678,21 +680,27 @@ extension SwiftLinkPreview {
         var image = image
 
         // TODO: account for HTML <base>
-        if let canonicalUrl = canonicalUrl, let finalUrl = finalUrl {
-            if finalUrl.hasPrefix("https:") {
+        if let canonicalUrl = canonicalUrl, let finalUrl = finalUrl, let proto = finalUrl.split(separator: ":").first {
+            if image.hasPrefix("/") {
                 if image.hasPrefix("//") {
-                    image = "https:" + image
-                } else if image.hasPrefix("/") {
-                    image = "https://" + canonicalUrl + image
+                    // image url is //domain/path
+                    image = proto + ":" + image
+                } else {
+                    // image url is /path relative to base url
+                    image = proto + "://" + canonicalUrl + image
                 }
-            } else if image.hasPrefix("//") {
-                image = "http:" + image
-            } else if image.hasPrefix("/") {
-                image = "http://" + canonicalUrl + image
+            } else if !image.contains("://") {
+                // image is relative to request url
+                let requestUrl = removeSuffixIfNeeded(finalUrl)
+                if requestUrl.hasSuffix("/") {
+                    image = requestUrl + image
+                } else {
+                    image = (requestUrl as NSString).deletingLastPathComponent + "/" + image
+                }
             }
         }
 
-        return removeSuffixIfNeeded(image)
+        return image
 
     }